Move clients and ANC containers to pages
