+++ src/App/App.tsx
@@ -15,8 +15,8 @@ import ConnectedHeader from '../containers/ConnectedHeader';
 
 import SideMenu from '../components/page/SideMenu';
 import { CLIENT_URL } from '../constants';
-import ConnectedANCListView from '../containers/ANC/list';
-import ConnectedClientList from '../containers/Clients/List';
+import ConnectedANCListView from '../containers/pages/ANC/list';
+import ConnectedClientList from '../containers/pages/Clients/List';
 import Home from '../containers/pages/Home/Home';
 import { oAuthUserInfoGetter } from '../helpers/utils';
 import './App.css';
